Merge-sort transformed to work with doubles .... It has to be converted to the qsort signature ....

+++ c/src/sort_utils.h
@@ -198,9 +198,9 @@ mergesort_array (int a[],
                  int temp[]);
 
 extern void
-sort_utils_mergesort (int a[],
+sort_utils_mergesort (double a[],
                       const size_t size,
-                      int temp[]);
+                      double temp[]);
 
 
 #endif /* SORT_UTILS_H */
